update docs, detypescriptify, simplify
