+++ docs/index.mdx
@@ -3,7 +3,7 @@ name: Home
 route: /
 ---
 
-import Hero from "./hero";
+import Hero from './hero';
 
 # `react-wordcloud`
