+++ docs/props/props.mdx
@@ -3,14 +3,12 @@ name: Props
 route: /props
 ---
 
-import { Props } from "docz";
+import { Props } from 'docz';
 
-import ReactWordcloud from "~/src";
-
-import CallbacksInterface from "./callbacks-interface";
-import ComponentProps from "./component-props";
-import OptionsInterface from "./options-interface";
-import WordInterface from "./word-interface";
+import CallbacksInterface from './callbacks-interface';
+import ComponentProps from './component-props';
+import OptionsInterface from './options-interface';
+import WordInterface from './word-interface';
 
 # Props
