+++ docs/usage/basic.mdx
@@ -4,11 +4,10 @@ menu: Usage
 route: /usage/basic
 ---
 
-import { Playground } from "docz";
+import { Playground } from 'docz';
 
-import ReactWordcloud from "~/src";
-
-import words from "../words";
+import ReactWordcloud from '../react-wordcloud';
+import words from '../words';
 
 # Basic Example
