+++ docs/usage/callbacks.mdx
@@ -4,12 +4,11 @@ menu: Usage
 route: /usage/callbacks
 ---
 
-import { select } from "d3-selection";
-import { Playground } from "docz";
+import { select } from 'd3-selection';
+import { Playground } from 'docz';
 
-import ReactWordcloud from "~/src";
-
-import words from "../words";
+import ReactWordcloud from '../react-wordcloud';
+import words from '../words';
 
 # Callbacks
 
@@ -25,9 +24,9 @@ In the example below, we can pass custom `getWordColor` and `getWordTooltip` cal
   <ReactWordcloud
     words={words}
     callbacks={{
-      getWordColor: ({ value }) => (value > 50 ? "blue" : "red"),
+      getWordColor: ({ value }) => (value > 50 ? 'blue' : 'red'),
       getWordTooltip: ({ text, value }) =>
-        `${text} (${value}) [${value > 50 ? "good" : "bad"}]`,
+        `${text} (${value}) [${value > 50 ? 'good' : 'bad'}]`,
     }}
   />
 </Playground>
@@ -39,26 +38,26 @@ You can pass callbacks such as `onWordClick`, `onWordMouseEnter` and `onWordMous
 <Playground>
   {() => {
     const getCallback = callbackName => (word, event) => {
-      const isActive = callbackName !== "onWordMouseOut";
+      const isActive = callbackName !== 'onWordMouseOut';
       const element = event.target;
       const text = select(element);
       text
-        .on("click", () => {
+        .on('click', () => {
           if (isActive) {
-            window.open(`https://duckduckgo.com/?q=${word.text}`, "_blank");
+            window.open(`https://duckduckgo.com/?q=${word.text}`, '_blank');
           }
         })
         .transition()
-        .attr("background", "white")
-        .attr("font-size", isActive ? "300%" : "100%")
-        .attr("text-decoration", isActive ? "underline" : "none");
+        .attr('background', 'white')
+        .attr('font-size', isActive ? '300%' : '100%')
+        .attr('text-decoration', isActive ? 'underline' : 'none');
     };
     return (
       <ReactWordcloud
         callbacks={{
-          onWordClick: getCallback("onWordClick"),
-          onWordMouseOut: getCallback("onWordMouseOut"),
-          onWordMouseOver: getCallback("onWordMouseOver"),
+          onWordClick: getCallback('onWordClick'),
+          onWordMouseOut: getCallback('onWordMouseOut'),
+          onWordMouseOver: getCallback('onWordMouseOver'),
         }}
         words={words}
       />
