+++ docs/usage/optimizations.mdx
@@ -4,8 +4,7 @@ menu: Usage
 route: /usage/optimizations
 ---
 
-import ReactWordcloud from '~/src';
-
+import ReactWordcloud from '../react-wordcloud';
 import words from '../words';
 
 # Optimizations
