+++ docs/usage/size.mdx
@@ -6,8 +6,7 @@ route: /usage/size
 
 import { Playground } from 'docz';
 
-import ReactWordcloud from '~/src';
-
+import ReactWordcloud from '../react-wordcloud';
 import words from '../words';
 
 # Size
