+++ docs/usage/transitions.mdx
@@ -7,8 +7,7 @@ route: /usage/transitions
 import { Playground } from 'docz';
 import { useState } from 'react';
 
-import ReactWordcloud from '~/src';
-
+import ReactWordcloud from '../react-wordcloud';
 import words from '../words';
 
 # Transitions
