+++ package.json
@@ -25,7 +25,7 @@
   "main": "dist/index.js",
   "module": "dist/index.module.js",
   "source": "index.js",
-  "typings": "src/index.d.ts",
+  "typings": "types/index.d.ts",
   "files": [
     "dist",
     "types/index.d.ts"
