--- src/hooks.d.ts
@@ -1,7 +1,0 @@
-import { MinMaxPair, Selection } from './types';
-
-export function useResponsiveSvgSelection<T>(
-  minSize: MinMaxPair,
-  initialSize?: MinMaxPair,
-  svgAttributes?: Record<string, string>,
-): [React.MutableRefObject<HTMLDivElement>, Selection, MinMaxPair];
