--- src/optimized-d3-cloud.d.ts
@@ -1,3 +1,0 @@
-import d3Cloud from 'd3-cloud';
-
-export default d3Cloud;
