+++ tsconfig.json
@@ -1,14 +1,10 @@
 {
   "compilerOptions": {
     "allowJs": true,
-    "baseUrl": ".",
     "checkJs": true,
     "esModuleInterop": true,
     "jsx": "react",
     "noEmit": true,
-    "paths": {
-      "~/*": ["./*"]
-    }
   },
   "exclude": ["**/dist", "**/node_modules"]
 }
